+++ examples/ssl_debug/ssl_debug.pas
@@ -5,23 +5,14 @@ program ssl_debug;
 // Use: lazbuild --build-mode=Debug ssl_debug.lpi
 
 uses
-  SysUtils, Request
-  {$IFDEF WINDOWS}, Windows{$ENDIF};
-
-{$IFDEF WINDOWS}
-function SetDllDirectoryW(lpPathName: LPCWSTR): BOOL; stdcall; external 'kernel32.dll';
-{$ENDIF}
+  SetDllPath,  // MUST be first to set DLL search path before Request unit loads OpenSSL
+               // Note: May not override System32 OpenSSL if it's in Known DLLs registry
+  SysUtils, Request;
 
 var
   Response: TResponse;
+
 begin
-  {$IFDEF WINDOWS}
-  // Force Windows to search the executable directory first for DLLs
-  // This prevents System32 DLLs from taking priority over local DLLs
-  SetDllDirectoryW(PWideChar(UnicodeString(ExtractFilePath(ParamStr(0)))));
-  {$ENDIF}
-
-
   WriteLn('=== OpenSSL Debug Information ===');
   WriteLn;